html/semantics/forms/event-propagate-disabled.html

<!DOCTYPE html>
<title>Event propagation on disabled form elements</title>
<meta charset=utf8>
<meta name=timeout content=long>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>

<div id=cases>
  <input> <!-- Santity check with non-disabled control -->
  <select disabled></select>
  <select disabled>
    <!-- <option> can't be clicked as it doesn't have its own painting area -->
    <option>foo</option>
  </select>
  <fieldset disabled>Text</fieldset>
  <fieldset disabled><span class="target">Span</span></fieldset>
  <button disabled>Text</button>
  <button disabled><span class="target">Span</span></button>
  <textarea disabled></textarea>
  <input disabled type="button">
  <input disabled type="checkbox">
  <input disabled type="color" value="#000000">
  <input disabled type="date">
  <input disabled type="datetime-local">
  <input disabled type="email">
  <input disabled type="file">
  <input disabled type="image">
  <input disabled type="month">
  <input disabled type="number">
  <input disabled type="password">
  <input disabled type="radio">
  <!-- Native click will click the bar -->
  <input disabled type="range" value="0">
  <!-- Native click will click the slider -->
  <input disabled type="range" value="50">
  <input disabled type="reset">
  <input disabled type="search">
  <input disabled type="submit">
  <input disabled type="tel">
  <input disabled type="text">
  <input disabled type="time">
  <input disabled type="url">
  <input disabled type="week">

  <!--
    WebDriver can't click on type="file"
    https://github.com/w3c/webdriver/issues/1666
  <input disabled="disabled" type="file">
  <input disabled="disabled" type="hidden"> -->
</div>

<script>
  /**
   * @param {Element} element
   */
  function getEventFiringTarget(element) {
    return element.querySelector(".target") || element;
  }

  const allEvents = ["pointermove", "mousemove", "pointerdown", "mousedown", "pointerup", "mouseup", "click"];

  /**
   * @param {*} t
   * @param {Element} element
   * @param {Element} observingElement
   */
  function setupTest(t, element, observingElement) {
    /** @type {string[]} */
    const observedEvents = [];
    const controller = new AbortController();
    const { signal } = controller;
    const listenerFn = t.step_func(event => observedEvents.push(event.type));
    for (const event of allEvents) {
      observingElement.addEventListener(event, listenerFn, { signal });
    }
    t.add_cleanup(() => controller.abort());

    const target = getEventFiringTarget(element);
    return { target, observedEvents };
  }

  /**
  * @param {object} options
  * @param {Element & { disabled: boolean }} options.element
  * @param {Element} options.observingElement
  * @param {string[]} options.expectedEvents
  * @param {(target: Element) => (Promise<void> | void)} options.clickerFn
  * @param {string} options.title
  */
  function promise_event_test({ element, observingElement, expectedEvents, nonDisabledExpectedEvents, clickerFn, title }) {
    promise_test(async t => {
      const { target, observedEvents } = setupTest(t, element, observingElement);

      await t.step_func(clickerFn)(target);
      await new Promise(resolve => t.step_timeout(resolve, 0));

      const expected = element.disabled ? expectedEvents : nonDisabledExpectedEvents;
      assert_array_equals(observedEvents, expected, "Observed events");

    }, `${title} on ${element.outerHTML}, observed from <${observingElement.localName}>`);
  }

  /**
   * @param {object} options
   * @param {Element & { disabled: boolean }} options.element
   * @param {string[]} options.expectedEvents
   * @param {(target: Element) => (Promise<void> | void)} options.clickerFn
   * @param {string} options.title
   */
  function promise_event_test_both({ element, expectedEvents, nonDisabledExpectedEvents, clickerFn, title }) {
    for (const observingElement of [element, document.body]) {
      promise_event_test({ element, observingElement, expectedEvents, nonDisabledExpectedEvents, clickerFn, title });
    }
  }

  function trusted_click(target) {
    return new test_driver.Actions()
      .addPointer("default-mouse", "mouse")
      .pointerMove(0, 0, target, 0)
      .pointerDown()
      .pointerUp()
      .send();
  }

  const mouseEvents = ["mousemove", "mousedown", "mouseup", "click"];
  const pointerEvents = ["pointermove", "pointerdown", "pointerup"];

  const elements = document.getElementById("cases").children;
  for (const element of elements) {
    promise_event_test({
      element,
      observingElement: element,
      expectedEvents: pointerEvents,
      nonDisabledExpectedEvents: allEvents,
      clickerFn: trusted_click,
      title: "Trusted click"
    })

    promise_event_test({
      element,
      observingElement: document.body,
      expectedEvents: allEvents,
      nonDisabledExpectedEvents: allEvents,
      clickerFn: trusted_click,
      title: "Trusted click"
    })

    const eventFirePair = [
      [MouseEvent, mouseEvents],
      [PointerEvent, pointerEvents]
    ];

    for (const [eventInterface, events] of eventFirePair) {
      promise_event_test_both({
        element,
        expectedEvents: events,
        nonDisabledExpectedEvents: events,
        clickerFn: target => {
          for (const event of events) {
            target.dispatchEvent(new eventInterface(event, { bubbles: true }))
          }
        },
        title: `Dispatch new ${eventInterface.name}()`
      })
    }

    promise_event_test_both({
      element,
      expectedEvents: getEventFiringTarget(element) === element ? [] : ["click"],
      nonDisabledExpectedEvents: ["click"],
      clickerFn: target => target.click(),
      title: `click()`
    })
  }
</script>
